Added Dashboard to admin, created logout api route
