+++ src/app/admin/dashboard/notice/delete-notice/page.js
