+++ src/app/api/auth/verify/route.js
@@ -9,7 +9,9 @@ export const GET = async (req) => {
       return new Response(JSON.stringify({ success: false, error: 'Method Not Allowed' }), { status: 405 });
     }
 
-    const token = await req.cookies._parsed.get('token').value;
+    // const token = await req.cookies._parsed.get('token').value;
+    const cookie = req.cookies?._parsed?.get('token');
+    const token = cookie?.value;
 
     if (!token) {
       return new Response(JSON.stringify({ success: false, message: 'Unauthorized! No token' }), { status: 401 });
